Small cleanup around generators
