+++ lib/src/generator/generator.dart
@@ -7,6 +7,8 @@ library;
 
 import 'package:dartdoc/src/dartdoc_options.dart';
 import 'package:dartdoc/src/generator/generator_backend.dart';
+import 'package:dartdoc/src/generator/html_generator_backend.dart';
+import 'package:dartdoc/src/generator/templates.dart';
 import 'package:dartdoc/src/logging.dart';
 import 'package:dartdoc/src/model/model.dart';
 import 'package:dartdoc/src/model_utils.dart';
@@ -305,4 +307,16 @@ List<DartdocOption> createGeneratorOptions(
             'Consider using if building many versions of the docs for public '
             'SEO; learn more at https://goo.gl/gktN6F.'),
   ];
+}
+
+/// Creates a [Generator] with an [HtmlGeneratorBackend] backend.
+Generator initHtmlGenerator(
+  DartdocGeneratorOptionContext context, {
+  required FileWriter writer,
+}) {
+  var templates = HtmlAotTemplates();
+  var options = DartdocGeneratorBackendOptions.fromContext(context);
+  var generatorBackend = HtmlGeneratorBackend(
+      options, templates, writer, context.resourceProvider);
+  return Generator(generatorBackend);
 }
